Check Item Service

+++ src/main/java/com/typecasters/apitowerofwords/Service/ItemService.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class ItemService {
@@ -17,79 +19,52 @@ public class ItemService {
     @Autowired
     private ItemRepository item_repo;
 
-        @Autowired
-    private UserRepository user_repo;
-
-    @Autowired
-    private UserItemRepository user_item_repo;
-
-
-
-
-      public ItemService(ItemRepository item_repo, UserRepository user_repo, UserItemRepository user_item_repo){
-            this.item_repo = item_repo;
-            this.user_repo = user_repo;
-            this.user_item_repo = user_item_repo;
+    // Create
+    public ItemEntity insertItem(ItemEntity item) {
+        return item_repo.save(item);
     }
 
-    public ItemService(ItemRepository item_repo){
-        this.item_repo = item_repo;
-    }
 
-    @Transactional
-    public ItemEntity addItemWithUserItems(ItemEntity item){
-        ItemEntity saved_item = item_repo.save(item);
-
-        List<UserEntity> users = user_repo.save(item);
-
-        for(UserEntity user : users){
-            UserItemEntity user_item = new UserItemEntity();
-            userItem.setItem(savedItem);
-            userItem.setUserDetails(user.getUserDetails());
-            user_item_repo.save(user_item);
-        }
-
-        return saved_item;
-    }
-
+    // Read
     public List<ItemEntity> getAllItems() {
         return item_repo.findAll();
     }
 
-    public ItemEntity getItembyId(int item_id){
-        return item_repo.findById(item_id).orElse(null);
+    public Optional<ItemEntity> getItemById(int item_id) {
+        return item_repo.findById(item_id);
     }
 
-    @Transactional
-    public ItemEntity updateItem(int item_id, ItemEntity updated_item){
-        ItemEntity existing_item = item_repo.findById(item_id).orElse(null);
+    // Update
+    @SuppressWarnings({"finally", "ReturnInsideFinallyBlock"})
+    public ItemEntity updateItem(int item_id, ItemEntity new_item_details){
+        ItemEntity item = new ItemEntity();
 
-        if  (existing_item != null){
+        try {
+            item = item_repo.findById(item_id).get();
 
-            if (updated_item.getItem_name() != null){
-                existing_item.setItem_name(updated_item.getItem_name());
-            }
+            item.setItem_name(new_item_details.getItem_name());
+            item.setItem_description(new_item_details.getItem_description());
+            item.setItem_price(new_item_details.getItem_price());
+            item.setImage_path(new_item_details.getImage_path());
+        }catch (NoSuchElementException ex) {
+            throw new NoSuchElementException("Item " + item_id + " does not exist!");
+        }finally {
+            return item_repo.save(item);
+        }
+    }
 
-            if  (updated_item.getImage_path() != null){
-                existing_item.setImage_path(updated_item.getImage_path());
-            }
 
-            if (updated_item.getItem_description() != null){
-                existing_item.setItem_description(updated_item.getItem_description());
-            }
+    // Delete
+    public String deleteItem(int item_id) {
+        String msg = "";
 
-            if (updated_item.getItem_price() != 0){
-                existing_item.setItem_price(updated_item.getItem_price());
-            }
+        if(item_repo.findById(item_id).isPresent()) {
+            item_repo.deleteById(item_id);
 
-            return item_repo.save(existing_item);
+            msg = "Item " + item_id + " is successfully deleted!";
         }
 
-        return null;
-    }
-
-    public void deleteItem(int item_id) {
-        item_repo.deleteById(item_id);
+        return msg;
     }
